+++ frontend/src/app/restaurant/[id]/page.tsx
@@ -2,8 +2,8 @@ import Image from 'next/image'
 import { notFound, redirect } from 'next/navigation'
 import { LuArrowUpRight } from 'react-icons/lu'
 import FavoriteButton from '~/components/features/restaurant/FavoriteButton'
+import ShareRestaurantDialog from '~/components/features/restaurant/share/ShareRestaurantDialog'
 import Section from '~/components/layout/Section'
-import ShareRestaurantDialog from '~/components/ui/dialogs/ShareRestaurantDialog'
 import { Table, TableBody, TableCell, TableHead, TableRow } from '~/components/ui/table'
 import { fetchRestaurantDetail } from '~/services/fetch-restaurant-detail'
 import { getGoogleMapsEmbedUrl } from '~/services/get-google-maps-embed-url'
